Added property tests to this project
Note: These tests are failing because the pack and unpack functions in
Data.ByteString.Char8 are for ASCII *only*. QuickCheck is easily finding
Unicode strings that don't convert correctly using this method. This
will be addressed in another commit.

Also added dist-* to .gitignore for more modern cabal build artifacts.