Changed Char8 un/pack to Text de/encode
If anything but ASCII or ISO-8859-1 is being worked with, it's safer to
convert between String and ByteString through Text's encodeUtf8 and
decodeUtf8 functions. The pack functions in Char8 are for ASCII only.
This may bite people more than they think, say with UTF-8, which is VERY
common.

This fixes the broken property tests added in an earlier commit.